+++ src/app/cart/models/cart.model.ts
@@ -1,5 +1,7 @@
-export interface ICartItem {
-  name: string;
-  quantity: number;
-  amount: number;
+export default class CartItemModel {
+  constructor(
+    public name: string,
+    public quantity: number,
+    public amount: number,
+  ) { }
 }