+++ src/app/cart/services/cart.service.ts
@@ -1,28 +1,54 @@
 import { Injectable } from '@angular/core';
-import { ICartItem } from '../models/cart.model';
+import ProductModel from 'src/app/products/models/product.model';
+import CartItemModel from '../models/cart.model';
 
 @Injectable({
   providedIn: 'root'
 })
 export class CartService {
-
-  public cartList: ICartItem[] = [];
-  public totalAmount: number = 0;
+  cartList: CartItemModel[] = [];
 
   constructor() { }
 
-  public getCartItems(): ICartItem[] {
+  public getCartItems(): CartItemModel[] {
     return this.cartList;
   }
 
-  public addItemToCartList(name: string, price: number): void {
-    const itemToAdd = this.cartList.find(item => item.name === name);
+  public addItemToCartList(product: ProductModel): void {
+    const itemToAdd = this.cartList.find(item => item.name === product.name);
     if (!itemToAdd) {
-      this.cartList.push({ name: name, quantity: 1, amount: price })
+      this.cartList.push({ name: product.name, quantity: 1, amount: product.price })
     } else {
-      const pricePerItem = itemToAdd.amount / itemToAdd.quantity;
       itemToAdd.quantity += 1;
-      itemToAdd.amount = Number((pricePerItem * itemToAdd.quantity).toFixed(2));
+      itemToAdd.amount = Number((product.price * itemToAdd.quantity));
     }
+  }
+
+  public getTotalQuantity(): number {
+    return Array.from(this.cartList.values()).reduce(
+      (acc, item) => acc + item.quantity,
+      0
+    );
+  }
+
+  public getTotalSum(): number {
+    return Array.from(this.cartList.values()).reduce(
+      (acc, item) => acc + item.amount,
+      0
+    );
+  }
+
+  public increaseQuantity(cartItem: CartItemModel): void {
+    cartItem.quantity++;
+  }
+  public decreaseQuantity(cartItem: CartItemModel): void {
+    if (cartItem.quantity === 1) {
+      this.deleteItem(cartItem);
+      return;
+    }
+    cartItem.quantity--;
+  }
+  public deleteItem(cartItem: CartItemModel): void {
+    this.cartList.splice(this.cartList.indexOf(cartItem), 1);
   }
 }