+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 import { ProductsService } from "../../services/products.service";
-import { IProductModel } from "../../models/product.model";
+import { CartService } from 'src/app/cart/services/cart.service';
+import ProductModel from '../../models/product.model';
 
 @Component({
   selector: 'app-product-list',
@@ -8,12 +9,19 @@ import { IProductModel } from "../../models/product.model";
   styleUrls: ['./product-list.component.css']
 })
 export class ProductListComponent implements OnInit {
-  public products: IProductModel[] = [];
 
-  constructor(private productsService: ProductsService) { }
+  @Output()
+  addProduct: EventEmitter<ProductModel> = new EventEmitter();
+
+  public products: ProductModel[] = [];
+
+  constructor(private productsService: ProductsService, private cartService: CartService) { }
 
   ngOnInit(): void {
     this.products = this.productsService.getProducts();
+  }
+  onAddToCart(product: ProductModel): void {
+    this.cartService.addItemToCartList(product);
   }
 
 }