+++ src/app/products/models/product.model.ts
@@ -1,14 +1,6 @@
 import { Category } from "./category.model";
 
-export interface IProductModel {
-  name: string;
-  description?: string;
-  price: number;
-  category: Category;
-  isAvailable: boolean;
-}
-
-/*export class Product {
+export default class ProductModel {
   constructor(
     public name: string,
     public description: string | null,
@@ -16,4 +8,12 @@ export interface IProductModel {
     public category: Category,
     public isAvailable: boolean = true,
   ) { }
-}*/
+
+  // export interface IProductModel {
+  //   name: string;
+  //   description?: string;
+  //   price: number;
+  //   category: Category;
+  //   isAvailable: boolean;
+  // }
+}