+++ CHANGELOG.md
@@ -12,3 +12,14 @@
 ### Refactored 
 - Realisation by functionality instead of by entities.
 
+## [1.0.1] - 2021-12-19
+### Added
+- CartItemComponent
+
+### Refactored 
+- Communication between ProductList and Product components using input/output
+- IProductModel replaced by Product
+- CartService
+- CartListComponent
+- @Inputs for for Product properties replaced by one single @Input for total object
+